+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - Various code refactoring
 - Listen to events for screen resolution changes and update overlays
     accordingly
+- Renamed "blockers" to "frame" and "opacity overlay" to "brightness filter"
+    for clarity
 
 #### 0.1.1